Implement first_valid_index for DataFrame & Integrate with Series
