+++ databricks/koalas/missing/frame.py
@@ -56,7 +56,6 @@ class _MissingPandasLikeDataFrame(object):
     eval = unsupported_function('eval')
     ewm = unsupported_function('ewm')
     first = unsupported_function('first')
-    first_valid_index = unsupported_function('first_valid_index')
     idxmax = unsupported_function('idxmax')
     idxmin = unsupported_function('idxmin')
     infer_objects = unsupported_function('infer_objects')
